# cmd/server/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: true
    type: object
  poll.Poll:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      options:
        items:
          $ref: '#/definitions/poll.PollOption'
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
  poll.PollOption:
    properties:
      option:
        type: string
      users:
        items:
          type: object
        type: array
    type: object
  user.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: This is the backend service documentation for the voting app
  license: {}
  title: Voting App API Documentation
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login endpoint that returns a new access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register endpoint that makes a new user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gin.H'
  /auth/verify:
    get:
      consumes:
      - application/json
      description: Verify's the user auth token and returns the user's data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
  /polls:
    get:
      consumes:
      - application/json
      description: Poll list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/poll.Poll'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
    post:
      consumes:
      - application/json
      description: Create poll
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/poll.Poll'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
  /users:
    get:
      consumes:
      - application/json
      description: Users list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
swagger: "2.0"